[update] 自动播放支持修改样式,分割同传样式修改器为components
